requirements.txt

Flask==3.0.3
python-dotenv==1.0.1
SQLAlchemy==2.0.32
Flask-Cors==4.0.1
psycopg2-binary==2.9.9
pandas==2.2.2
openpyxl==3.1.5
qrcode==7.4.2
Pillow==10.4.0
